requirements.txt
aioredis==1.3.0
amqp==2.5.2
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.6
django-background-tasks==1.2.0
django-celery-beat==1.5.0
django-compat==1.0.15
django-cors==0.1
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-redis==4.10.0
django-timezone-field==3.1
django-user-agents==0.4.0
djangorestframework==3.10.3
docopt==0.6.2
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.20.0
gunicorn==19.9.0
h2==3.1.1
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.6
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
numpy==1.17.4
Pillow==6.2.1
pipreqs==0.4.10
priority==1.3.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
PyYAML==5.1.2
ratelim==0.1.6
redis==3.2.0
requests==2.22.0
reverse-geocoder==1.5.1
scipy==1.3.2
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
supervisor==4.1.0
Twisted==19.7.0
txaio==18.8.1
ua-parser==0.8.0
urllib3==1.25.6
user-agents==2.0
vine==1.3.0
whitenoise==4.1.4
yarg==0.1.9
zipp==0.6.0
zope.interface==4.6.0
